test(anyharness): split worker tailer tests
